+++ src/constants.rs
@@ -1,1 +1,4 @@
+use macroquad::prelude::{vec2, Vec2};
+
 pub const MOVEMENT_SPEED: f32 = 200.0;
+pub const WINDOWS_SIZE: Vec2 = vec2(370.0, 320.0);
